Struct asynchronous::EventLoopEmitter

pub struct EventLoopEmitter<Ev> {
    // some fields omitted
}

Event Emitter wrapper that can be cloned to pass between threads

Methods

impl<Ev> EventLoopEmitter<Ev>

fn emit(&self, event: Ev) -> Result<(), Ev>

Triggers an event "Ev" once. Returns the same event if the event loop is not active.

fn is_active(&self) -> bool

Returns true if the event loop is running

fn finish(self) -> EventLoopEmitter<Ev>

Finishes the event loop immediately.